renamed particular to item and moved repository nipype interfaces into separate module
