+++ arcana/dataset/__init__.py
@@ -1,4 +1,4 @@
-from .particular import Dataset, Field
+from .item import Dataset, Field
 from .collection import DatasetCollection, FieldCollection
 from .spec import DatasetSpec, FieldSpec, BaseSpec
 from .base import BaseField, BaseDataset, BaseDatum
